+++ lib/field_validators/dynamic_select_field.js
@@ -1,13 +1,14 @@
 'use strict';
 var registry = require('component-registry').globalRegistry
 var createObjectPrototype = require('component-registry').createObjectPrototype;
-var BaseField = require('./base_fields').BaseField;
+var SelectField = require('./select_field');
 var i18n = require('../utils').i18n;
 var Promise = require('es6-promise')
 
 /*
 
-    A Select Field allows you to choose a single value from a selection
+    A Dynamic Select Field allows you to choose a single value from a selection
+    that is created at runtime
 
     valueType: TextField,
     options: [{name: "1", title: "First"}]
@@ -16,12 +17,12 @@ var Promise = require('es6-promise')
 
 // TODO: Write tests
 
-var ISelectField = require('../interfaces').ISelectField;
+var IDynamicSelectField = require('../interfaces').IDynamicSelectField;
 
-var SelectField = createObjectPrototype({
-    implements: [ISelectField],
+var DynamicSelectField = createObjectPrototype({
+    implements: [IDynamicSelectField],
 
-    extends: [BaseField],
+    extends: [SelectField],
     
     constructor: function (options) {
         this._IBaseField.constructor.call(this, options);
@@ -36,6 +37,7 @@ var SelectField = createObjectPrototype({
     },
     
     validate: function (inp, options, context, async) {
+        // Check that this isn't undefined if it is required
         var error = this._IBaseField.validate.call(this, inp);
         if (error) { return error };
     
@@ -54,7 +56,34 @@ var SelectField = createObjectPrototype({
             return
         };
         
-        var options = this.options
+        try {
+            var util = registry.getUtility(this.options.utilityInterface, this.options.name)
+        } catch (err) {
+            throw new Error("DynamicSelectField can't find utility (" + this.options.utilityInterface.name + ", " + this.options.name + ")")
+        }
+        var options = util.getOptions(inp, options, context)
+        if (!async && options.then) {
+            throw new Error('DynamicSelectField got a promise when fetching options from util, but validation wasn\'t async!')
+        } else if (options.then) {
+            return options.then(function (options) {
+                var matches = false
+                for (var i = 0; i < options.length; i++) {
+                    if (options[i].name === inp) {
+                        matches = true
+                        break
+                    }
+                }
+                if (matches) {
+                    return Promise.resolve(undefined)
+                } else {
+                    return Promise.resolve({
+                        type: 'constraint_error',
+                        i18nLabel: i18n('isomorphic-schema--select_field_value_error', 'The selected value is not allowed'),
+                        message: "Valt värde finns inte i listan över tillåtna värden"
+                    })
+                }
+            })
+        }
 
         var matches = false
         for (var i = 0; i < options.length; i++) {
@@ -84,10 +113,10 @@ var SelectField = createObjectPrototype({
     },
 
     getOptionTitle: function (inp) {
-        for (var i=0; i < this.options.length; i++) {
-            if (this.options[i].name === inp) return this.options[i].title 
-        }
+        var util = registry.getUtility(this.options.utilityInterface, this.options.name, undefined)
+        // We will either get a title or a promise, returning either way
+        return util.getOptionTitle(inp)
     }
 });
 
-module.exports = SelectField;
+module.exports = DynamicSelectField;
