+++ lib/field_validators/index.js
@@ -5,6 +5,7 @@ var HTMLAreaField = require('./html_area_field');
 var EmailField = require('./email_field');
 var CreditCardField = require('./credit_card_field');
 var SelectField = require('./select_field');
+var DynamicSelectField = require('./dynamic_select_field');
 var MultiSelectField = require('./multi_select_field');
 var ListField = require('./list_field');
 var ObjectField = require('./object_field');
@@ -31,6 +32,7 @@ module.exports = {
     creditCardField: function (options) {return new CreditCardField(options)},
     orgNrField: function (options) {return new OrgNrField(options)},
     selectField: function (options) {return new SelectField(options)},
+    dynamicSelectField: function (options) {return new DynamicSelectField(options)},
     multiSelectField: function (options) {return new MultiSelectField(options)},
     listField: function (options) {return new ListField(options)},
     objectField: function (options) {return new ObjectField(options)},
