+++ lib/field_validators/multi_select_field.js
@@ -59,40 +59,6 @@ var MultiSelectField = createObjectPrototype({
 
             var options = this.options
 
-            if (!Array.isArray(this.options) && this.options.utilityInterface) {
-                try {
-                    var util = registry.getUtility(this.options.utilityInterface, this.options.name)
-                } catch (err) {
-                    throw new Error("MultiSeletField can't find utility (" + this.options.utilityInterface.name + ", " + this.options.name + ")")
-                }
-                // TODO: Handle if we get a Promise from getOptions!
-                var options = util.getOptions(inp, options, context)
-                if (!async && options.then) {
-                    throw new Error('MultiSeletField got a promise when fetching options from util, but validation wasn\'t async!')
-                } else if (options.then) {
-                    return options.then(function (options) {
-                        var matches = false
-                        for (var i = 0; i < options.length; i++) {
-                            if (options[i].name === inp) {
-                                matches = true
-                                break
-                            }
-                        }
-                        if (matches) {
-                            return Promise.resolve(undefined)
-                        } else {
-                            return Promise.resolve({
-                                type: 'constraint_error',
-                                i18nLabel: i18n('isomorphic-schema--multi_select_field_value_error', 'One or more of the selected values is not allowed'),
-                                message: "Ett eller flera valda värden finns inte i listan över tillåtna värden"
-                            })
-                        }
-                    })
-                }
-            } else {
-                options = this.options
-            }
-
             // Check that selected value is in options list
             var matches = false
             for (var i = 0; i < options.length; i++) {
