+++ lib/index.js
@@ -18,7 +18,7 @@ module.exports.fieldObjectPrototypes = {
     EmailField: require('./field_validators/email_field'),
     CreditCardField: require('./field_validators/credit_card_field'),
     SelectField: require('./field_validators/select_field'),
-    MultiSelectField: require('./field_validators/multi_select_field'),
+    DynamicSelectField: require('./field_validators/dynamic_select_field'),
     ListField: require('./field_validators/list_field'),
     ObjectField: require('./field_validators/object_field'),
     ObjectRelationField: require('./field_validators/object_relation_field'),
