+++ lib/interfaces.js
@@ -70,8 +70,16 @@ module.exports.ISelectField = createInterface({
     name: 'ISelectField'
 });
 
+module.exports.IDynamicSelectField = createInterface({
+    name: 'IDynamicSelectField'
+});
+
 module.exports.IMultiSelectField = createInterface({
     name: 'IMultiSelectField'
+});
+
+module.exports.IDynamicMultiSelectField = createInterface({
+    name: 'IDynamicMultiSelectField'
 });
 
 /* Richtext HTML field */
